requirements.txt
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
imageio==2.22.3
kiwisolver==1.4.4
matplotlib==3.6.2
nd2reader==3.3.0
networkx==2.8.8
numpy==1.23.4
packaging==21.3
Pillow==9.3.0
PIMS==0.6.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
scikit-image==0.19.3
scipy==1.9.3
six==1.16.0
slicerator==1.1.0
tifffile==2022.10.10
xmltodict==0.13.0
